Implement config for cfg atoms and refactor lsp configs modules
