+++ crates/engine/project/src/project/update/package.rs
@@ -196,7 +196,7 @@ enum RebuildResidency {
 
 pub(crate) fn rebuild_resident_from_source(state: &mut ProjectState) -> anyhow::Result<()> {
     let workspace = state.workspace.clone();
-    let workspace_lowering_config = state.workspace_lowering_config;
+    let workspace_lowering_config = state.workspace_lowering_config.clone();
     let cargo_metadata_config = state.cargo_metadata_config.clone();
     let body_ir_policy = state.body_ir_policy;
     let indexing_preference = state.indexing_preference;
