+++ editors/code/README.md
@@ -73,6 +73,8 @@ just client lint
   "rust-glancer.server.path": null,
   "rust-glancer.server.extraEnv": {},
   "rust-glancer.cargo.target": null,
+  "rust-glancer.cfg.test": true,
+  "rust-glancer.cfg.atoms": [],
   "rust-glancer.cache.packageResidency": "workspace-and-path-deps",
   "rust-glancer.trace.server": "off",
   "rust-glancer.diagnostics.onStartup": false,
